Fetch a runner's real Sverigelistan through their own session
externalLoginUrl mints a one-time link for a member of the organisation
whose key we hold; following it with a cookie jar yields that person's
session, and their ranking page comes back without the fee wall. No
password is involved anywhere — the link needs only a personId, and that
is the same id the ranking pages use.

Two boundaries, both measured: a session cannot be minted for anyone
outside our organisation (403, whichever club id is passed), but any
signed-in session can read every runner's page. Sverigelistan is a
subscription, not a per-person permission, so borrowing one member's
session for other people's data is a setting that has to be switched on
deliberately rather than something that happens.

Points keep their two decimals. Rounding to whole points threw away the
hundredths that separate one place from the next, and it hid a view bug:
"resultat i snittet" sorted by points descending and so listed the worst
results, because Sverigelistan counts downwards.

+++ samples/Orientera.Domain/Domain/Ranking.cs
@@ -8,7 +8,7 @@ public sealed record RankingResult
     public required CompetitionId Competition { get; init; }
     public required string CompetitionName { get; init; }
     public required DateOnly Date { get; init; }
-    public required int Points { get; init; }
+    public required double Points { get; init; }
 
     /// <summary>Whether this result is one of the six that make up the current average.</summary>
     public required bool IsCounting { get; init; }
@@ -25,14 +25,18 @@ public sealed record RankingSnapshot
 {
     public required PersonId Person { get; init; }
     public required DateOnly Date { get; init; }
-    public required int Points { get; init; }
+    public required double Points { get; init; }
     public required int NationalPlace { get; init; }
 
     /// <summary>Points change since the previous snapshot.</summary>
-    public required int Trend { get; init; }
+    public required double Trend { get; init; }
 
-    public IReadOnlyDictionary<Discipline, int> DisciplinePoints { get; init; } =
-        new Dictionary<Discipline, int>();
+    /// <summary>
+    /// Sverigelistan is published to two decimals and the differences are small — a runner can sit
+    /// a few hundredths from the place above. Rounding to whole points threw that away.
+    /// </summary>
+    public IReadOnlyDictionary<Discipline, double> DisciplinePoints { get; init; } =
+        new Dictionary<Discipline, double>();
 
     public required IReadOnlyList<RankingResult> Results { get; init; }
